[form] replace chart using iframe
