+++ static/gwml2/css/main.css
@@ -1033,3 +1033,8 @@ input[type="checkbox"] {
     white-space: pre-wrap;
     text-align: left;
 }
+
+.many-to-many iframe {
+    width: 100%;
+    height: 530px;
+}
