+++ static/gwml2/js/measurement-chart.js
@@ -1,35 +1,90 @@
-function renderMeasurementChart($chart, id, data, xLabel, yLabel, parameters, parameter_to, units, unit_to) {
-    $chart.html(`<canvas id="${id}_chart_canvas"></canvas>`);
+const chartColors = [
+    "rgb(255, 99, 132)",
+    "rgb(54, 162, 235)",
+    "rgb(153, 102, 255)",
+    "rgb(255, 205, 86)",
+    "rgb(75, 192, 192)",
+    "rgb(255, 159, 64)",
+    "rgb(201, 203, 207)"
+]
 
-    // render parameters and unit selection
-    {
-        let html = '<select class="parameter-selection">';
-        parameters.forEach(function (param) {
-            if (parameter_to === param) {
-                html += `<option value="${param}" selected>` + param + '</option>'
-            } else {
-                html += `<option value="${param}">` + param + '</option>'
-            }
-        })
-        html += '</select>';
-        html += '<select class="unit-selection">';
-        units.forEach(function (unit) {
-            if (unit_to === unit) {
-                html += `<option value="${unit}" selected>` + unit + '</option>'
-            } else {
-                html += `<option value="${unit}">` + unit + '</option>'
-            }
-        })
-        html += '</select><br><br><br>';
-        $chart.prepend(html);
-        $chart.find('select').change(function () {
-            chartFunctions[id](
-                $chart.find('.parameter-selection').val(),
-                $chart.find('.unit-selection').val()
-            )
-        })
-    }
-    var ctx = document.getElementById(`${id}_chart_canvas`).getContext("2d");
+function convertMeasurementData(input, unit_to, parameter_to, top_borehole_elevation, ground_surface_elevation) {
+    let data = {};
+    input.map(function (row) {
+        let parameter = row['par'];
+        let value = row['v'];
+        let methodology = row['mt'];
+        let unit = row['u'];
+        let time = row['dt'];
+        value = unitConvert(unit, unit_to, value);
+
+        switch (parameter_to) {
+            case 'Water level elevation a.m.s.l.':
+                switch (parameter) {
+                    case 'Water depth [from the top of the well]':
+                        value = top_borehole_elevation - value;
+                        break;
+                    case 'Water depth [from the ground surface]':
+                        value = ground_surface_elevation - value;
+                        break;
+                    default:
+                        value = 0 - value;
+                        break;
+                }
+                parameter = parameter_to;
+                break;
+
+            case 'Water depth [from the ground surface]':
+                switch (parameter) {
+                    case 'Water depth [from the top of the well]':
+                        value = (top_borehole_elevation - ground_surface_elevation) - value;
+                        break;
+                    case 'Water level elevation a.m.s.l':
+                        value = ground_surface_elevation - value;
+                        break;
+                    default:
+                        value = 0 - value;
+                        break;
+                }
+                parameter = parameter_to;
+                break;
+
+            case 'Water depth [from the top of the well]':
+                switch (parameter) {
+                    case 'Water depth [from the ground surface]':
+                        value = 0 - value - (top_borehole_elevation - ground_surface_elevation);
+                        break;
+                    case 'Water level elevation a.m.s.l':
+                        value = top_borehole_elevation - value;
+                        break;
+                    default:
+                        value = 0 - value;
+                        break;
+                }
+                parameter = parameter_to;
+                break;
+        }
+
+
+        if (!data[parameter]) {
+            data[parameter] = []
+        }
+        if (parameter === parameter_to) {
+            data[parameter].push({
+                t: new Date(time * 1000),
+                y: value,
+                methodology: methodology,
+                unit: unit_to
+            });
+        }
+    })
+    return data;
+}
+
+let chart = null;
+
+function renderMeasurementChart(identifier, data, xLabel, yLabel) {
+    let ctx = document.getElementById(`${identifier}-chart`).getContext("2d");
     let dataset = [];
     let idx = 0;
     $.each(data, function (key, value) {
@@ -44,7 +99,7 @@ function renderMeasurementChart($chart, id, data, xLabel, yLabel, parameters, pa
         })
         idx += 1;
     });
-    new Chart(ctx, {
+    const options = {
         type: 'line',
         data: {
             datasets: dataset
@@ -73,12 +128,8 @@ function renderMeasurementChart($chart, id, data, xLabel, yLabel, parameters, pa
                 intersect: true,
                 callbacks: {
                     label: function (tooltipItem, allData) {
-                        let label = allData.datasets[tooltipItem.datasetIndex].label || '';
                         let data = allData.datasets[tooltipItem.datasetIndex].data[tooltipItem.index];
-                        if (label) {
-                            label += ' : ';
-                        }
-                        label += tooltipItem.yLabel;
+                        let label = tooltipItem.yLabel + ' ' + data.unit;
                         if (data.methodology) {
                             label += ', methodology :' + data.methodology;
                         }
@@ -86,6 +137,13 @@ function renderMeasurementChart($chart, id, data, xLabel, yLabel, parameters, pa
                     }
                 }
             }
-        },
-    });
+        }
+    };
+    if (!chart) {
+        chart = new Chart(ctx, options);
+    } else {
+        chart.data = options.data;
+        chart.options = options.options;
+        chart.update();
+    }
 }
