--- templates/groundwater_form/_many-to-many-graph.html
@@ -1,137 +1,0 @@
-<div id="{{ id }}_chart" class="chart measurement-chart">
-</div>
-<script>
-    chartFunctions['{{ id }}'] = function (parameter_to, unit_to) {
-        {# get top and surface of borehole #}
-        let top_borehole_elevation = null;
-        let ground_surface_elevation = null;
-
-        let parameters = [];
-        let units = [];
-
-        {# format data #}
-        let data = {};
-        $('#{{ id }}_table table tr').each(function (index) {
-            if (index > 0) {
-                const $units = $(this).find('*[name ="value_unit"] option').not("[hidden=hidden]");
-                const time = $(this).find('*[name ="time"]').val();
-                const methodology = $(this).find('*[name ="methodology"]').val();
-                let parameter = $(this).find('*[name ="parameter"] option:selected').text();
-
-                {# if parameter to is empty, assign in as parameter #}
-                if (!parameter_to) {
-                    parameter_to = parameter
-                }
-
-                {# get parameters options #}
-                {# if parameters is empty and it is level measurement #}
-                {# put all parameters as option #}
-                {# otherwise just put current parameter into parameters #}
-                if (parameters.length === 0 && [
-                    'Water level elevation a.m.s.l.',
-                    'Water depth [from the top of the well]',
-                    'Water depth [from the ground surface]'
-                ].includes(parameter)) {
-                    $(this).find('*[name ="parameter"] option').each(function () {
-                        if ($(this).attr('value')) {
-                            parameters.push($(this).text());
-                        }
-                    });
-                }
-                if (parameter && !parameters.includes(parameter)) {
-                    parameters.push(parameter);
-                }
-
-                {# check unit_to if unit is empty#}
-                unit_to = unit_to !== undefined ? unit_to : $($units[0]).text();
-
-                let value = getQuantityInputValue($(this).find('.quantity-input'), unit_to);
-
-                {# skip if value is empty #}
-                if (value === undefined) {
-                    return
-                }
-                {# check top and ground surfaces #}
-                if (top_borehole_elevation === null) {
-                    const wellElevation = getTopGroundSurfaceWell(unit_to)
-                    top_borehole_elevation = wellElevation['top_borehole'];
-                    ground_surface_elevation = wellElevation['ground_surface'];
-                }
-
-                {# check the parameter of level #}
-                switch (parameter_to) {
-                    case 'Water level elevation a.m.s.l.':
-                        switch (parameter) {
-                            case 'Water depth [from the top of the well]':
-                                value = top_borehole_elevation - value;
-                                break;
-                            case 'Water depth [from the ground surface]':
-                                value = ground_surface_elevation - value;
-                                break;
-                            default:
-                                value = 0 - value;
-                                break;
-                        }
-                        parameter = parameter_to;
-                        break;
-
-                    case 'Water depth [from the ground surface]':
-                        switch (parameter) {
-                            case 'Water depth [from the top of the well]':
-                                value = (top_borehole_elevation - ground_surface_elevation) - value;
-                                break;
-                            case 'Water level elevation a.m.s.l':
-                                value = ground_surface_elevation - value;
-                                break;
-                            default:
-                                value = 0 - value;
-                                break;
-                        }
-                        parameter = parameter_to;
-                        break;
-
-                    case 'Water depth [from the top of the well]':
-                        switch (parameter) {
-                            case 'Water depth [from the ground surface]':
-                                value = 0 - value - (top_borehole_elevation - ground_surface_elevation);
-                                break;
-                            case 'Water level elevation a.m.s.l':
-                                value = top_borehole_elevation - value;
-                                break;
-                            default:
-                                value = 0 - value;
-                                break;
-                        }
-                        parameter = parameter_to;
-                        break;
-                }
-
-                {# assign data#}
-                if (!data[parameter]) {
-                    data[parameter] = []
-                }
-                if (parameter === parameter_to) {
-                    data[parameter].push({
-                        t: new Date(time),
-                        y: value,
-                        methodology: methodology
-                    });
-                    {# get units list #}
-                    if (units.length === 0) {
-                        $units.each(function () {
-                            if ($(this).attr('value')) {
-                                units.push($(this).text());
-                            }
-                        });
-                    }
-                }
-            }
-        });
-
-        {# create chart #}
-        renderMeasurementChart(
-            $('#{{ id }}_chart'), '{{ id }}', data, 'Time',
-            parameter_to, parameters, parameter_to, units, unit_to
-        )
-    }
-</script>
