+++ templates/groundwater_form/_many-to-many.html
@@ -1,9 +1,10 @@
+{% load gwml2_forms %}
 <div id="{{ id }}" class="many-to-many">
     <h4>{{ title }}
         {% if help_text %}
             <i class="fa fa-question-circle-o" aria-hidden="true" data-toggle="tooltip" title="{{ help_text }}"></i>
         {% endif %}
-        {% if use_chart %}
+        {% if use_chart and theform.url_chart %}
             <i class="fa fa-line-chart chart-table-toggle" aria-hidden="true"
                onclick="chartTableToggle(this,'{{ id }}')"
             ></i>
@@ -13,7 +14,15 @@
         ></i>
     </h4>
     {% include "groundwater_form/_many-to-many-table.html" %}
-    {% if use_chart %}
-        {% include "groundwater_form/_many-to-many-graph.html" %}
+    {% if use_chart and theform.url_chart %}
+        <div id="{{ id }}_chart" class="chart measurement-chart">
+            <iframe src="{{ theform.url_chart }}" style="border: none"></iframe>
+            <i style="display: inline-block;
+                color: #888;
+                font-size: 12px;
+                margin-top: 10px;">
+                This chart is coming from the database. New / updated data that are in current form needs to be saved to make it show on chart.***
+            </i>
+        </div>
     {% endif %}
 </div>
