+++ templates/groundwater_form/main.html
@@ -148,7 +148,6 @@
     <script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-beta.1/dist/js/select2.min.js"></script>
 
     <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
-    <script src="{% static 'gwml2/js/measurement-chart.js' %}"></script>
     <script src="{% static 'gwml2/js/unit-converter.js' %}"></script>
     <script src="{% static 'gwml2/js/utilities.js' %}"></script>
     <script src="{% static 'gwml2/js/scripts.js' %}"></script>
