+++ templates/plugins/measurements_chart.html
@@ -6,7 +6,7 @@
     <script src="{% static "lib/js/moment-with-locales.js" %}"></script>
     <script src="{% static 'js/jquery/3.5.1/jquery.min.js' %}"></script>
     <script src="{% static 'gwml2/js/unit-converter.js' %}"></script>
-    <script src="{% static 'gwml2/js/measurements_chart.js' %}"></script>
+    <script src="{% static 'gwml2/js/measurement-chart.js' %}"></script>
     <script src="{% static "js/chartjs/2.8.0/chartjs.min.js" %}"></script>
 </head>
 <body>
